추가 : Sanctum 설정 및 인증 로직

+++ blog/app/Http/Controllers/CommentController.php
@@ -16,12 +16,12 @@ class CommentController extends Controller
             return abort( 404 );
         }
 
-        $author = $request->input('author');
+        $user = $request->user();
         $content = $request->input('content');
 
         $comment = new Comment();
+        $comment->user_id = $user->id;
         $comment->post_id = $post->id;
-        $comment->author = $author;
         $comment->content = $content;
         $comment->save();
 
@@ -32,6 +32,11 @@ class CommentController extends Controller
         $comment = Comment::where( 'post_id', $postId )->where( 'id', $id )->first();
 
         if( !$comment ) abort( 404 );
+
+        $user = $request->user();
+        if( $user->id !== $comment->user_id ) {
+            return abort( 403 );
+        }
 
         $comment->delete();
